+++ src/collections/Menu/MenuHeader.d.ts
@@ -1,0 +1,21 @@
+import * as React from 'react';
+
+interface MenuHeaderProps {
+  [key: string]: any;
+
+  /** An element type to render as (string or function). */
+  as?: any;
+
+  /** Primary content. */
+  children?: React.ReactNode;
+
+  /** Additional classes. */
+  className?: string;
+
+  /** Shorthand for primary content. */
+  content?: React.ReactNode;
+}
+
+const MenuHeader: React.ComponentClass<MenuHeaderProps>;
+
+export default MenuHeader;
