+++ src/collections/Message/MessageContent.d.ts
@@ -1,0 +1,19 @@
+import * as React from 'react';
+
+interface MessageContentProps {
+  [key: string]: any;
+
+  /** An element type to render as (string or function). */
+  as?: any;
+
+  /** Primary content. */
+  children?: React.ReactNode;
+
+  /** Additional classes. */
+  className?: string;
+}
+
+const MessageContent: React.StatelessComponent<MessageContentProps>;
+
+export default MessageContent;
+
