+++ src/collections/Table/TableFooter.d.ts
@@ -1,0 +1,9 @@
+import * as React from 'react';
+import { TableHeaderProps } from './TableHeader';
+
+interface TableFooterProps extends TableHeaderProps {
+}
+
+const TableFooter: React.StatelessComponent<TableFooterProps>;
+
+export default TableFooter;
