+++ src/collections/Table/TableHeaderCell.d.ts
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import { TableCellProps } from './TableCell';
+
+interface TableHeaderCellProps extends TableCellProps {
+  sorted?: 'ascending' | 'descending';
+}
+
+const TableHeaderCell: React.StatelessComponent<TableHeaderCellProps>;
+
+export default TableHeaderCell;
